requirements.txt
glfw==2.8.0
imgui==2.0.0
numpy==2.2.1
pillow==11.1.0
PyOpenGL==3.1.7
